requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
idna==3.10
MarkupSafe==3.0.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
